Sort out new commands from master, delete commands.cpp
:crab: :crab:
