+++ src/commands/area.cpp
@@ -283,3 +283,9 @@ void AOClient::cmdJudgeLog(int argc, QStringList argv)
         sendServerMessage(filteredmessage);
     }
 }
+
+void AOClient::cmdAfk(int argc, QStringList argv)
+{
+    is_afk = true;
+    sendServerMessage("You are now AFK.");
+}
